Support both UTF-8 and Windows-1252 encoded files
If we end up trying to insert in text that isn't UTF-8, then we try
through a bit of a hack to convert it into some sort of UTF-8-ish
encoding before inserting in the text equivalent.

Alternative solutions were to just drop the invalid characters, or
replace them with "?".

If this is causing you problems, first check that you are correctly
loading and parsing the input string in a valid encoding. Make sure that
`text.valid_encoding?` returns true.

Beforei this change, this gem only supported UTF-8-encoded files.

Copied example spec from html2text (PHP) component.